extracted method to base class to remove duplicate code
